document GET email endpoint
This commit documents the GET email API endpoint in the `openapi.yaml` file and contains some formatting of the file as well as changes to descriptions, summaries and the introduction. In the introduction I've documented how to set the environment variables because adding them to the compose file as it said here before doesn't work.